requirements/base.txt
python-dotenv==0.19.2
py-cord==1.7.3
asyncpraw==7.5.0
requests==2.27.1
html2text==2020.1.16
tabulate==0.8.9